annotator/depth/__init__.py
import torch
import requests
from PIL import Image
from diffusers import StableDiffusionDepth2ImgPipeline

pipe = StableDiffusionDepth2ImgPipeline.from_pretrained(
   "stabilityai/stable-diffusion-2-depth",
   torch_dtype=torch.float16,
).to("cuda")

class DepthDetector:
    def __call__(self, img):
        img = Image.fromarray(img)
        img = pipe(img)
        return img